test: move osakit to a separate test with no harness
to force it run on the main thread
