+++ Cargo.toml
@@ -29,6 +29,11 @@ env_logger = "0.10.0"
 tempfile = "3.8.0"
 defer = "0.2.1"
 
+[[test]]
+name = "osakit"
+path = "tests/osakit.rs"
+harness = false
+
 
 [target.'cfg(target_os = "macos")'.dependencies]
 objc2 = "0.5.1"
